Refactor; add to README.md.
