+++ package.json
@@ -4,17 +4,17 @@
   "private": true,
   "dependencies": {
     "lodash": "^4.17.15",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-router": "^5.0.1",
-    "react-router-dom": "^5.0.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.1.1",
     "react-spring": "^8.0.27",
-    "react-transition-group": "^4.2.2",
+    "react-transition-group": "^4.3.0",
     "react-use-gesture": "^5.2.4",
-    "react-vertical-timeline-component": "^2.4.0",
+    "react-vertical-timeline-component": "^2.5.0",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.88.0"
+    "semantic-ui-react": "^0.88.2"
   },
   "scripts": {
     "start": "react-scripts start",
